UI for admin and user

+++ src/main/java/pills/service/PillServiceImpl.java
@@ -7,7 +7,9 @@ import javax.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import inti.ws.spring.exception.client.BadRequestException;
 import pills.dao.PillDao;
+import pills.entity.Category;
 import pills.entity.Pill;
 import pills.models.AddPillModel;
 import pills.models.PillModel;
@@ -28,7 +30,10 @@ public class PillServiceImpl implements PillService{
 		pillDao.save(pill);
 		return mapUtility.mapPill(pill);
 	}
-	public void deletePill(Pill pill){
+	public void deletePill(Integer id)throws BadRequestException{
+		if (id <= 0)
+			throw new BadRequestException("Required parameters are either missing or invalid");
+		Pill pill = new Pill(id);
 		pillDao.delete(pill);
 	}
 	public PillModel viewPill(Integer id){
